+++ CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
+### Added
+- Can set global locale instance used by error instances with `Errors::setGlobalLocale()`.
+
 ### Changed
 - Relax `icanboogie/inflector` version dependency.